Reorder builds to improve build times
The main issue is that all iOS builds are in a block and there are only 5
simultaneous builds allowed by our Travis CI plan. So if there's a backlog
of iOS builders, then the entire build process stalls. So instead of having
them all clumped together, distribute the Mac and iOS builds throughout the
build pipeline.