Fix test_thread_signal* test failures with Cargo 1.33.0
Previous versions of Cargo would create and destroy a new thread for
each test.  Cargo 1.33.0 instead creates a thread pool and reuses the
same thread for multiple tests.  Some Nix tests that changed the
per-thread sigmask began to fail as a result, because they didn't do any
cleanup.

The easiest solution is to spawn a new thread for each of those tests.