Disable test-aio-drop in GNU environments
This test occasionally fails in Travis on x86_64-unknown-linux-gnu.  The
failure takes the form of the test executable receiving a random signal.
Sometimes SIGHUP, sometimes SIGKILL, etc.  I think this must be a bug in
glibc, even though I can't reproduce it in my development environment.
But it interferes with CI too much to leave enabled.