Fix fork test and enable doc test
Replaces `println!`  with raw `libc::write`. `println!` isn't guaranteed
to be async-signal-safe, and almost certainly *isn't* due to internal
buffering and locking.

Adds a call to `libc::_exit` in the child arm, so that it doesn't fall
through and start executing the parent code.

Adds a call to `waitpid` in the parent arm, to clean up the child
process.

Removes the `no_run` directive, so that it's run in the doc tests.