UnixAddr: replace path_len with sun_len
Within UnixAddr, replace the path_len variable (length of the sun_path
field) with sun_len (length of the whole structure).  This is more
similar to how other sockaddr types work, and it's the same way that the
BSDs use the sun_len field.

Also, don't require that sun_path be nul-terminated. The OS doesn't
require it.